requirements.txt
llama-cpp-python==0.2.24
rich>=13.5.2